Moved T1Analysis to a separate file (#130)
