+++ qiskit_experiments/characterization/__init__.py
@@ -37,7 +37,8 @@ Analysis
     T2StarAnalysis
     SpectroscopyAnalysis
 """
-from .t1 import T1, T1Analysis
+from .t1 import T1
+from .t1_analysis import T1Analysis
 from .qubit_spectroscopy import QubitSpectroscopy, SpectroscopyAnalysis
 from .ef_spectroscopy import EFSpectroscopy
 from .t2star_experiment import T2StarExperiment, T2StarAnalysis
